+++ app/controllers/replies_controller.rb
@@ -1,8 +1,11 @@
 class RepliesController < ApplicationController
     get '/posts/:post_id/replies/new' do 
-        if logged_in?
-            @post = Post.find_by_id(params[:post_id])
+        post = Post.find_by_id(params[:post_id])
+        if logged_in? && post
+            @post = post
             erb :'/replies/new'
+        elsif logged_in?
+            redirect '/posts'
         else
             @error = "Please log in"
             erb :'/users/login'
@@ -11,11 +14,13 @@ class RepliesController < ApplicationController
 
     post '/posts/:post_id/replies' do 
         post = Post.find_by_id(params[:post_id])
-        if !params[:reply][:content].empty?
+        if post && !params[:reply][:content].empty?
             reply = current_user.replies.create(params[:reply])
             reply.post = post
             reply.save
             redirect "/posts/#{post.id}"
+        elsif post
+            redirect '/posts'
         else
             @error = "Your reply can't be empty"
             erb :'/replies/new'
@@ -24,9 +29,12 @@ class RepliesController < ApplicationController
 
 
     get '/posts/:post_id/replies/:id/edit' do 
-        if logged_in?
-            @reply = Reply.find_by_id(params[:id])
+        reply = Reply.find_by_id(params[:id])
+        if logged_in? && reply
+            @reply = reply
             erb :'/replies/edit'
+        elsif logged_in?
+            redirect '/posts'
         else
             @error = "Please log in"
             erb :'/users/login'
@@ -35,23 +43,27 @@ class RepliesController < ApplicationController
 
     delete '/replies/:id' do 
         reply = Reply.find_by_id(params[:id])
-        post_id = reply.post.id
-        if reply.user == current_user
+        if reply && reply.user == current_user
+            post_id = reply.post.id
             reply.destroy
             redirect "/posts/#{post_id}"
+        elsif reply
+            redirect "/posts/#{reply.post.id}"
         else
-            redirect "/posts/#{post_id}"
+            redirect "/posts"
         end
-    end 
+    end
 
     patch '/replies/:id' do
         reply = Reply.find_by_id(params[:id])
-        post_id = reply.post.id
-        if reply.user == current_user
+        if reply && reply.user == current_user
+            post_id = reply.post.id
             reply.update(params[:reply])
             redirect "/posts/#{post_id}"
+        elsif reply
+            redirect "/posts/#{reply.post.id}" 
         else
-            redirect "/posts/#{post_id}" 
+            redirect '/posts'
         end
     end
 end
